requirements.txt
# apipkg==1.5
# atomicwrites==1.4.0
# attrs==19.3.0
# blinker==1.4
# click==8.0.4
# colorama==0.4.3
# distlib==0.3.4
# execnet==1.7.1
# filelock==3.6.0
Flask==2.0.3
# Flask-Mail==0.9.1
# flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.5.1
# greenlet==1.1.2
idna
# importlib-metadata==1.6.0
# inflect==4.1.0
# itsdangerous==2.1.1
# Jinja2==3.0.3
# Mako==1.2.0
# MarkupSafe==2.1.1
# marshmallow==3.15.0
# more-itertools==8.2.0
mysqlclient==2.1.0
# packaging==20.3
# pep8==1.7.1
# platformdirs==2.5.1
# pluggy==0.13.1
# py==1.8.1
pyinstaller
pymssql==2.2.4
PyMySQL==0.9.3
pyodbc==4.0.32
# pyparsing==2.4.7
# pytest==5.4.2
# pytest-cache==1.0
# pytest-pep8==1.0.6
python-dotenv
# six==1.14.0
SQLAlchemy==1.3.17
# sqlparse==0.4.2
# sqltap==0.3.11
urllib3==1.26.9
virtualenv==20.13.4
# wcwidth==0.1.9
Werkzeug==2.0.3
XlsxWriter==3.0.3
# zipp==3.1.0